+++ source/others/f64_to_i64.c
@@ -84,10 +84,10 @@ f64_to_i64(float64_t a, uint8_t roundingMode, bool exact)
             }
             sig <<= -shiftDist;
             extSig[indexWord(3, 2)] = sig >> 32;
-            extSig[indexWord(3, 1)] = sig;
+            extSig[indexWord(3, 1)] = (uint32_t)sig;
         } else {
             extSig[indexWord(3, 2)] = sig >> 32;
-            extSig[indexWord(3, 1)] = sig;
+            extSig[indexWord(3, 1)] = (uint32_t)sig;
             softfloat_shiftRightJam96M(extSig, shiftDist, extSig);
         }
         return softfloat_roundPackMToI64(sign, extSig, roundingMode, exact);
